requirements.txt
# ============================================
# Core ML frameworks (Intel + RTX 5060 优化)
# ============================================
torch==2.9.1
torchvision==0.24.1
ultralytics>=8.3.0  # YOLO 目标检测
# 新版 transformers 对应的预编译 tokenizers 轮子支持 Python 3.14（abi3），BLIP 依赖改为可选
transformers==4.57.5
tokenizers==0.22.2

# Intel Extension for PyTorch (CPU 优化)
# Intel Extension for PyTorch (仅 Linux，暂不支持当前 Windows/Python 3.14 环境)
intel-extension-for-pytorch>=2.4.0; platform_system == "Linux" and python_version < "3.13"

# NVIDIA CUDA 优化 (RTX 5060)
# 注意: 需要安装对应 CUDA 版本的 PyTorch
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124

# ============================================
# BLIP 模型 (优化版本)
# 说明：salesforce-lavis 目前与 Python 3.14/Windows 组合兼容性较差，默认跳过。
# 如需 BLIP，请在 Linux + Python<=3.12 环境下手动安装。
# ============================================
salesforce-lavis>=1.0.2; platform_system != "Windows" and python_version < "3.13"
accelerate>=0.33.0  # 混合精度和分布式训练
bitsandbytes>=0.43.0  # 量化支持，减少显存

# CLIP 模型
open-clip-torch>=2.24.0

# PEFT for LoRA fine-tuning
peft>=0.12.0

# ============================================
# 通义千问 VL API (对照组)
# ============================================
dashscope>=1.20.0  # 阿里云 DashScope SDK
openai>=1.40.0  # 兼容 OpenAI 格式调用

# ============================================
# 数据处理与 OpenCV 数据生成
# ============================================
pandas>=2.2.0
numpy>=1.26.0
Pillow>=10.4.0
datasets>=2.20.0
opencv-python>=4.10.0  # OpenCV 核心
opencv-contrib-python>=4.10.0  # OpenCV 扩展模块

# 图像增强
albumentations>=1.4.0
imgaug>=0.4.0

# ============================================
# 可视化
# ============================================
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.22.0
scikit-learn>=1.5.0  # for t-SNE
streamlit>=1.39.0

# ============================================
# 工具库
# ============================================
tqdm>=4.66.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
requests>=2.32.0
httpx>=0.27.0  # 异步 HTTP 客户端
aiohttp>=3.10.0  # 异步请求

# ============================================
# Jupyter 支持
# ============================================
jupyter>=1.0.0
ipywidgets>=8.1.0
notebook>=7.2.0

# ============================================
# 实验追踪 (可选)
# ============================================
# wandb>=0.17.0
# tensorboard>=2.17.0

# ============================================
# 开发工具
# ============================================
pytest>=8.3.0
ruff>=0.5.0
